# crates/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Shelfy API",
    "description": "HTTP API of Shelfy Web. JSON in camelCase, timestamps in unix milliseconds. Errors are `application/problem+json` documents whose `code` is stable; clients map it to their own messages.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "1"
  },
  "paths": {
...
    "/api/v1/extension/config": {
      "get": {
        "tags": ["extension"],
        "summary": "The extension's configuration: the minimum version, the kill switches of\neach platform's capture modes, pacing and stop thresholds (contract C3).",
        "description": "Conditional: send the `ETag` back in `If-None-Match` and an unchanged\nconfiguration answers 304. A change made by the operator (`admin flags`)\nshows within 30 seconds. This is the one route an outdated extension\nstill reaches: every other answers it 426 `extension_outdated`.",
        "operationId": "getExtensionConfig",
        "parameters": [
          {
            "name": "If-None-Match",
            "in": "header",
            "description": "The `ETag` of an earlier response. When the view has not changed since,\nthe answer is 304 with no body.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Shelfy-Extension",
            "in": "header",
            "description": "The extension's manifest version (`0.2.0`). An extension token's\nrequest without it, or below `minVersion`, gets 426\n`extension_outdated`, except `GET /extension/config`.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The configuration.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "`private, no-cache`."
              },
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Weak ETag of this configuration."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtensionConfig"
                }
              }
            }
          },
          "304": {
            "description": "Unchanged since the ETag in `If-None-Match`; no body.",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "The same ETag."
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "security": [
          {
            "bearer": ["ingest"]
          }
        ]
      }
    },
    "/api/v1/extension/pair": {
      "post": {
        "tags": ["extension"],
        "summary": "Exchanges a pairing code for the extension's token.",
        "description": "The extension calls it with the code that the web app got from `POST\n/me/tokens/pairing-code`, within 60 seconds, once. No cookie and no CSRF\nheaders are needed, and none is read. 400 `invalid_pairing_code` for a\ncode that is malformed, unknown, used or expired; 422 for a malformed\n`installId`, `label` or `version`; 426 `extension_outdated` for a version\nbelow `minVersion` (the code stays usable); 409 `conflict` when the\naccount holds 50 working tokens. Limit: the sign-in limit, 10 requests a\nminute per client over this route and the `/auth` routes (429).",
        "operationId": "pairExtension",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtensionPairRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Paired: the token, shown once.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairedExtension"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "security": [{}]
      }
    },
    "/api/v1/extension/status": {
      "get": {
        "tags": ["extension"],
        "summary": "Whether the account's browser extension is connected: one of its\nextension tokens made a request in the last 10 minutes, with the version\nit sent. The same object as the `extension.status` event, which follows\nevery change. The server keeps it in memory: after a restart it is\ndisconnected until the extension's next request.",
        "operationId": "getExtensionStatus",
        "responses": {
          "200": {
            "description": "The extension's status.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtensionStatusEvent"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
    "/api/v1/me/tokens/pairing-code": {
      "post": {
        "tags": ["account"],
        "summary": "Creates a code that pairs the browser extension with this account.",
        "description": "The web app hands the code to the extension (`chrome.runtime.sendMessage`),\nwhich exchanges it for its token at `POST /extension/pair` within 60\nseconds, once. Needs a sign-in or a re-authentication from the last 5\nminutes (403 `reauth_required` otherwise). 429 `rate_limited` while the\naccount holds 10 unused codes that have not expired.",
        "operationId": "createPairingCode",
        "responses": {
          "201": {
            "description": "The code and its expiry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairingCode"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorCode": {
        "type": "string",
        "description": "Stable, machine-readable error codes. Each one has a fixed HTTP status.\n\nNew codes may be added within `/api/v1`; existing ones never change meaning.",
        "enum": [
          "bad_request",
          "invalid_cursor",
          "invalid_link",
          "challenge_expired",
          "passkey_invalid",
          "invalid_device_code",
          "invalid_pairing_code",
          "unauthorized",
          "forbidden",
          "csrf_failed",
          "reauth_required",
          "quota_exceeded",
          "not_found",
          "method_not_allowed",
          "conflict",
          "upload_consumed",
          "payload_too_large",
          "unsupported_media_type",
          "validation_failed",
          "provider_key_invalid",
          "capture_blocked",
          "not_available",
          "user_locked",
          "extension_outdated",
          "rate_limited",
          "internal",
          "storage_full",
          "unavailable",
          "timeout"
        ]
      },
...
      "ExtensionConfig": {
        "type": "object",
        "description": "What `GET /api/v1/extension/config` answers: the extension's minimum\nversion, kill switches, pacing and stop thresholds (contract C3). Data\nonly; the extension refreshes it, and applies a kill switch within one\nrefresh.",
        "required": [
          "minVersion",
          "platforms",
          "maxSteps",
          "maxRunMs",
          "taskPollMinutes",
          "refreshPerSession"
        ],
        "properties": {
          "maxRunMs": {
            "type": "integer",
            "format": "int64",
            "description": "Longest sync run, ms.",
            "minimum": 0
          },
          "maxSteps": {
            "type": "integer",
            "format": "int32",
            "description": "Most steps (scrolls, replay pages) of one sync run.",
            "minimum": 0
          },
          "minVersion": {
            "type": "string",
            "description": "The oldest extension version the API serves: below it, every other\nrequest answers 426 `extension_outdated`."
          },
          "platforms": {
            "$ref": "#/components/schemas/ExtensionPlatforms",
            "description": "Per platform: kill switches, pacing, stop thresholds."
          },
          "refreshPerSession": {
            "type": "integer",
            "format": "int32",
            "description": "Most Instagram posts refreshed per IG-tab session.",
            "minimum": 0
          },
          "taskPollMinutes": {
            "type": "integer",
            "format": "int32",
            "description": "Minutes between two polls of `GET /ingest/tasks`.",
            "minimum": 0
          }
        }
      },
      "ExtensionInstagram": {
        "type": "object",
        "description": "Instagram's switches and pacing: X and Pinterest's, plus the REST\nreplay.",
        "required": [
          "passive",
          "replay",
          "scroll",
          "stopAfterKnown",
          "replayGapMs",
          "replayMaxPages",
          "scrollSettleMs"
        ],
        "properties": {
          "passive": {
            "type": "boolean",
            "description": "Passive capture is on (kill switch)."
          },
          "replay": {
            "type": "boolean",
            "description": "The REST replay is on (kill switch)."
          },
          "replayGapMs": {
            "type": "integer",
            "format": "int32",
            "description": "Pause between two replay pages, ms.",
            "minimum": 0
          },
          "replayMaxPages": {
            "type": "integer",
            "format": "int32",
            "description": "Most replay pages per run.",
            "minimum": 0
          },
          "scroll": {
            "type": "boolean",
            "description": "The scroll is on (kill switch)."
          },
          "scrollSettleMs": {
            "type": "integer",
            "format": "int32",
            "description": "Wait after each scroll step, ms.",
            "minimum": 0
          },
          "stopAfterKnown": {
            "type": "integer",
            "format": "int32",
            "description": "An incremental run stops at the first page boundary where this many\nknown items came in a row.",
            "minimum": 0
          }
        }
      },
      "ExtensionPairRequest": {
        "type": "object",
        "description": "Body of `POST /api/v1/extension/pair` (contract C2). Fields a later\nextension adds are ignored.",
        "required": ["code", "installId", "version"],
        "properties": {
          "code": {
            "type": "string",
            "description": "The pairing code the web app handed over (43 characters)."
          },
          "installId": {
            "type": "string",
            "description": "A random id the extension keeps while it is installed (16 to 128\nletters, digits, `-` or `_`). Pairing the same installation again\nrevokes the token it held."
          },
          "label": {
            "type": "string",
            "description": "A name for the account's token list, such as \"Chrome on macOS\"; up\nto 64 characters, blank is none."
          },
          "version": {
            "type": "string",
            "description": "The extension's manifest version (`0.2.0`); below `minVersion`, 426."
          }
        }
      },
      "ExtensionPlatform": {
        "type": "object",
        "description": "A platform's switches and pacing (X, Pinterest).",
        "required": ["passive", "scroll", "stopAfterKnown", "scrollSettleMs"],
        "properties": {
          "passive": {
            "type": "boolean",
            "description": "Passive capture is on (kill switch)."
          },
          "scroll": {
            "type": "boolean",
            "description": "The scroll is on (kill switch)."
          },
          "scrollSettleMs": {
            "type": "integer",
            "format": "int32",
            "description": "Wait after each scroll step, ms.",
            "minimum": 0
          },
          "stopAfterKnown": {
            "type": "integer",
            "format": "int32",
            "description": "An incremental run stops at the first page boundary where this many\nknown items came in a row.",
            "minimum": 0
          }
        }
      },
      "ExtensionPlatforms": {
        "type": "object",
        "description": "The per-platform part of [`ExtensionConfig`].",
        "required": ["instagram", "twitter", "pinterest"],
        "properties": {
          "instagram": {
            "$ref": "#/components/schemas/ExtensionInstagram",
            "description": "Instagram."
          },
          "pinterest": {
            "$ref": "#/components/schemas/ExtensionPlatform",
            "description": "Pinterest."
          },
          "twitter": {
            "$ref": "#/components/schemas/ExtensionPlatform",
            "description": "X."
          }
        }
      },
      "ExtensionStatusEvent": {
        "type": "object",
        "description": "`extension.status`: whether the user's browser extension is connected\n(plan §2.10, contract C8), sent when `connected` or `version` changes.\n`GET /api/v1/extension/status` answers the same object.",
        "required": ["connected", "version", "lastSeenAt"],
        "properties": {
          "connected": {
            "type": "boolean",
            "description": "One of the user's extension tokens made a request in the last 10\nminutes."
          },
          "lastSeenAt": {
            "type": ["integer", "null"],
            "format": "int64",
            "description": "The newest request of the user's extension since the server started,\nunix ms; `null` when there was none."
          },
          "version": {
            "type": ["string", "null"],
            "description": "The extension's version from its `X-Shelfy-Extension` header: the\nhighest among the connected browsers, or the last one known once\ndisconnected. `null` when none sent a valid version."
          }
        }
      },
      "FieldError": {
        "type": "object",
        "description": "One field that failed validation.",
        "required": ["field", "reason"],
        "properties": {
          "field": {
            "type": "string",
            "description": "The field, as named in the request (camelCase)."
          },
          "reason": {
            "type": "string",
            "description": "Why the value was refused, for developers."
          }
        }
      },
...
      "PairedExtension": {
        "type": "object",
        "description": "A paired extension's token.",
        "required": ["token", "tokenId", "scopes"],
        "properties": {
          "scopes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TokenScope"
            },
            "description": "What it may do: `ingest`, `tasks`, `uploads`, `lookup`."
          },
          "token": {
            "type": "string",
            "description": "The `extension` token (`shx_…`) for `Authorization: Bearer`. Shown\nonly now; keep it where page scripts cannot read it."
          },
          "tokenId": {
            "type": "string",
            "description": "Its id, as the account's token list names it."
          }
        }
      },
      "PairingCode": {
        "type": "object",
        "description": "A code that pairs the browser extension (contract C2).",
        "required": ["code", "expiresAt"],
        "properties": {
          "code": {
            "type": "string",
            "description": "The code (43 characters), for the extension's `POST /extension/pair`.\nHand it to the extension only; it works once."
          },
          "expiresAt": {
            "type": "integer",
            "format": "int64",
            "description": "When it stops working, unix ms: 60 seconds from now."
          }
        }
      },
...
      "Problem": {
        "type": "object",
        "description": "An error body: RFC 9457 problem details plus the stable `code`.",
        "required": ["type", "title", "status", "code"],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode",
            "description": "What went wrong. Clients map it to their own message."
          },
          "detail": {
            "type": "string",
            "description": "Developer-facing detail. Never UI prose, user content or secrets."
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "description": "The fields that failed validation (`validation_failed`)."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "The HTTP status of the response.",
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "The reason phrase of `status`. Not meant for display."
          },
          "type": {
            "type": "string",
            "description": "Always `about:blank`: `code` carries the problem type."
          }
        }
      },
...
      "TokenScope": {
        "type": "string",
        "description": "What a token may do (§2.9).",
        "enum": ["ingest", "tasks", "uploads", "lookup", "links:create", "migrate"]
      },
...
    },
    "responses": {
      "Problem": {
        "description": "An error, as RFC 9457 problem details with a stable `code`.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "shx_",
        "description": "A scoped API token (`shx_…`) of the extension, the iOS Shortcut or the migration CLI. Tokens cannot call routes that only accept the session."
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-shelfy_session",
        "description": "The web app's session, set by signing in. Requests that change state (POST, PUT, PATCH, DELETE) with this cookie must also send `X-Shelfy-Client: web` and an `Origin` equal to the public URL; otherwise they fail with 403 `csrf_failed`. A request with an `Authorization` header is never authenticated by this cookie."
      }
    }
  },
  "security": [
    {
      "session": []
    }
  ],
...
}